Move catkin clean test to correct file
